[SS-2018-015] Update jQuery version used in templates
jQuery v1 is mighty old at this point, and with the recent updates
elsewhere and the deprecation of entwine (which relies on old versions)
this provides us an excellent opportunity to update. Existing
functionality has been tested to remain working even with the bump of two
major versions. This is chiefly the Highlighter (on the search page
template) and the cursor key navigable main menu.

Also includes a minor template alteration to support the same change in
silverstripe/comments