FIX Navigation no longer uses flexbox, fix search form width and jumbotron margin on mobile
Bootstrap 4 uses flexbox for the navigation component. We have customised it
quite a bit for accessiblity, so this commit removes the flex attributes
and continues to use block display properties.

It is build mobile first, so responsive viewport modifications are at
the bottom of the nav.scss file. Also removed an unused click handler
which was preventing an enter key from toggling to close an open dropdown.