FIX shift to 8 + 3 grid system
As per design update - there was already a grid system in place, however
it was for Bootstrap 3, and the Bootstrap 4 grid has changed. Most notably
a new break size was added at the smaller end of the scale, pushing all
others "up one", so `md` became `lg`. This commit adds more consistency
throughout all templates, and conforms to newer design tweaks for Watea
(which uses Starter as a base via SS4 theme cascades).